getOriginalDataPointReservedYears sorted and NationalClasses minor component refactor
