--- src/client/pages/OriginalDataPoint/components/NationalClasses/NationalClass/index.ts
@@ -1,1 +1,0 @@
-export { default } from './NationalClass'
